+++ workspace/packages/components/siyuan/list/list.ts
@@ -15,12 +15,14 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import deepmerge from "deepmerge";
+/* 列表项 */
+export interface IListItem {
+    text: string, // 内容
+    icon?: string, // 图标 (Unicode 字符 / svg ID)
+    src?: string, // 图标 (图片 URL)
+    meta?: string, // 元信息
+    style?: string, // 样式
 
-export function merge(...args: any[]): Object {
-    return deepmerge.all(args);
-}
-
-export function mergeIgnoreArray(...args: any[]): Object {
-    return deepmerge.all(args, { arrayMerge: (_target, source, _options) => source });
+    fold?: boolean, // 是否折叠
+    children?: IListItem[], // 下级列表
 }