+++ workspace/packages/components/siyuan/setting/item/Input.svelte
@@ -32,6 +32,7 @@
     export let placeholder: string = ""; // Use it if type is text/number/textarea
     export let options: IOptions = []; // Use it if type is select
     export let limits: ILimits = { min: 0, max: 100, step: 1 }; // Use it if type is number/slider
+    export let height: number = 0; // Use it if type is textarea
 
     const dispatch = createEventDispatcher();
 
@@ -131,6 +132,7 @@
         class="b3-text-field"
         class:fn__block={block}
         class:fn__size200={!block && normal}
+        style:height={(height > 0) ? `${height}px` : undefined}
         {placeholder}
         bind:value={settingValue}
         on:change={changed}